WeChatSpiders/requirements.txt
scrapy
redis
pymysql
python-dateutil